pyproject.toml
[tool.poetry]
name = "formflow"
version = "0.1.0"
description = ""
authors = ["hamid <xhamid0011@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = ">=5.0,<5.1"
django-restframework = "^0.0.1"
celery = "^5.4.0"
redis = "^5.0.8"
django-getenv = "^1.3.2"
django-rest-swagger = "^2.2.0"
django-celery-beat = "^2.6.0"
django-celery-results = "^2.5.1"
drf-yasg = "^1.21.7"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
flower = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
